Taking the second step to Jisho
Jisho, as described in the website, will be a flat-file dictionary with the base being Reo's custom definition system. The implementation of the CDEF system was the first step and this is the second step, to create something that automatically generates the flat file definitions. It is basically a fork of the current Reo with everything stripped off.

Signed-off-by: mufeed20 <mufeed.ali53@gmail.com>