Fixing a UI bug and cleaning code
The revert that was made the other day has now been reverted. That is, a
revert has been reverted. This is due to recent changes in Gtk that have
caused bugs and warnings to rise in the system that was in use until now.
The code has also been given a slight upgrade. Only the syscheck function
has been improved in this commit. Its cleaner with less lines of code,
adheres to standards in variable declaration and therefore, might be
faster.

Signed-off-by: lastweakness <lastweakness@tuta.io>